create base page & remove navigation
